Predict opponent moves
To get around the problem of not knowing what the opponent does, we predict their moves and play as if that is on the board. This should be modified in future ideally. We run into a problem in the environment, I think it allows two placements on the same cell which shouldn't be possible. We also have a problem when our prediction leads to another required prediction